move api add/edit into api/events
